# docs/swagger/admin-api.yml
openapi: 3.1.0
info:
  title: Admin API
  description: Library of Coffee - Admin API
  contact:
    name: Library of Coffee
  version: 1.0.0
servers:
- url: /
  description: Current Server
security:
- Bearer: []
...
paths:
...
  /api/admin/plans/{id}:
    get:
      tags:
      - Plan
      summary: プラン詳細取得
      description: 指定されたIDのプランの詳細情報を取得します。
      operationId: getPlan
      parameters:
      - name: id
        in: path
        description: プランID
        required: true
        schema:
          type: string
        example: 00000000-0000-4000-8000-000000000024
      responses:
        "200":
          description: 取得成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlanDetailResponse"
              examples:
                success:
                  summary: 取得成功例
                  description: success
                  value:
                    id: 00000000-0000-4000-8000-000000000024
                    shopifyPlanId: test-plan-001
                    label: 定番
                    gramWeight: 60
                    beanQuantity: 4
                    price: 3800
                    type: SUBSCRIPTION
                    isRecommended: true
        "404":
          description: プランが見つかりません
          content:
            '*/*':
              examples:
                not_found:
                  summary: プランが見つからない場合
                  description: not_found
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 404
                    error: Not Found
                    path: /api/admin/plans/00000000-0000-4000-8000-000000000999
    put:
      tags:
      - Plan
      summary: プラン編集
      description: 指定されたIDのプランを編集します。全項目を置換します。
      operationId: updatePlan
      parameters:
      - name: id
        in: path
        description: プランID
        required: true
        schema:
          type: string
        example: 00000000-0000-4000-8000-000000000024
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
        required: true
      responses:
        "200":
          description: 編集成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlanResponse"
              examples:
                success:
                  summary: 編集成功例
                  description: success
                  value:
                    id: 00000000-0000-4000-8000-000000000024
        "400":
          description: バリデーションエラー
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad_request:
                  summary: バリデーションエラーの場合
                  description: bad_request
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 400
                    error: Bad Request
                    path: /api/admin/plans/00000000-0000-4000-8000-000000000024
        "404":
          description: プランが見つかりません
          content:
            '*/*':
              examples:
                not_found:
                  summary: プランが見つからない場合
                  description: not_found
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 404
                    error: Not Found
                    path: /api/admin/plans/00000000-0000-4000-8000-000000000999
        "409":
          description: ShopifyプランIDが重複しています
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                conflict:
                  summary: ShopifyプランIDが重複している場合
                  description: conflict
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 409
                    error: Conflict
                    path: /api/admin/plans/00000000-0000-4000-8000-000000000024
...
  /api/admin/plans:
    get:
      tags:
      - Plan
      summary: プラン一覧取得
      description: プランの一覧をページネーション付きで取得します。プラン表示名による部分一致検索が可能です。
      operationId: listPlans
      parameters:
      - name: page
        in: query
        description: ページ番号（0始まり）
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: size
        in: query
        description: 1ページあたりの件数
        required: false
        schema:
          type: integer
          format: int32
          default: 20
        example: 20
      - name: keyword
        in: query
        description: プラン表示名（部分一致検索）
        required: false
        schema:
          type: string
        example: 定番
      responses:
        "200":
          description: 取得成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlanListResponse"
              examples:
                success:
                  summary: 取得成功例
                  description: success
                  value:
                    items:
                    - id: 00000000-0000-4000-8000-000000000024
                      shopifyPlanId: test-plan-001
                      label: 定番
                      gramWeight: 60
                      beanQuantity: 4
                      price: 3800
                      type: SUBSCRIPTION
                      isRecommended: true
                    totalCount: 1
                    page: 0
                    size: 20
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: エラーレスポンス
      properties:
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
          example: 2026-02-23T12:00:00.000+00:00
        status:
          type: integer
          format: int32
          description: HTTPステータスコード
          example: 404
        error:
          type: string
          description: エラー概要
          example: Not Found
        message:
          type: string
          description: エラーメッセージ
          example: tastes must not be empty
        path:
          type: string
          description: リクエストパス
          example: /api/admin/resources/00000000-0000-4000-8000-000000000099
      required:
      - error
      - path
      - status
      - timestamp
    UpdatePlanRequest:
      type: object
      description: プラン編集リクエスト
      properties:
        shopifyPlanId:
          type: string
          description: ShopifyのプランID
          example: test-plan-001
        label:
          type: string
          description: プラン表示名
          example: 定番
        gramWeight:
          type: integer
          format: int32
          description: 1種あたりのグラム数（30 / 60 / 90）
          example: 60
        beanQuantity:
          type: integer
          format: int32
          description: 豆の種類数（3 / 4 / 5）
          example: 4
        price:
          type: integer
          format: int32
          description: 価格
          example: 3800
        type:
          type: string
          description: プラン種別（SUBSCRIPTION / SINGLE）
          example: SUBSCRIPTION
        isRecommended:
          type: boolean
          description: おすすめバッジ
          example: true
      required:
      - beanQuantity
      - gramWeight
      - isRecommended
      - label
      - price
      - shopifyPlanId
      - type
    PlanResponse:
      type: object
      description: プラン編集レスポンス
      properties:
        id:
          type: string
          description: プランID
          example: 00000000-0000-4000-8000-000000000024
      required:
      - id
...
    PlanListResponse:
      type: object
      description: プラン一覧レスポンス
      properties:
        items:
          type: array
          description: アイテム一覧
          items:
            $ref: "#/components/schemas/PlanSummaryResponse"
        totalCount:
          type: integer
          format: int64
          description: 全件数
          example: 42
        page:
          type: integer
          format: int32
          description: 現在のページ番号（0始まり）
          example: 0
        size:
          type: integer
          format: int32
          description: 1ページあたりの件数
          example: 20
      required:
      - items
      - page
      - size
      - totalCount
    PlanSummaryResponse:
      type: object
      description: プラン一覧アイテム
      properties:
        id:
          type: string
          description: プランID
          example: 00000000-0000-4000-8000-000000000024
        shopifyPlanId:
          type: string
          description: ShopifyのプランID
          example: test-plan-001
        label:
          type: string
          description: プラン表示名
          example: 定番
        gramWeight:
          type: integer
          format: int32
          description: 1種あたりのグラム数
          example: 60
        beanQuantity:
          type: integer
          format: int32
          description: 豆の種類数
          example: 4
        price:
          type: integer
          format: int32
          description: 価格
          example: 3800
        type:
          type: string
          description: プラン種別（SUBSCRIPTION / SINGLE）
          example: SUBSCRIPTION
        isRecommended:
          type: boolean
          description: おすすめバッジ
          example: true
      required:
      - beanQuantity
      - gramWeight
      - id
      - isRecommended
      - label
      - price
      - shopifyPlanId
      - type
    PlanDetailResponse:
      type: object
      description: プラン詳細レスポンス
      properties:
        id:
          type: string
          description: プランID
          example: 00000000-0000-4000-8000-000000000024
        shopifyPlanId:
          type: string
          description: ShopifyのプランID
          example: test-plan-001
        label:
          type: string
          description: プラン表示名
          example: 定番
        gramWeight:
          type: integer
          format: int32
          description: 1種あたりのグラム数
          example: 60
        beanQuantity:
          type: integer
          format: int32
          description: 豆の種類数
          example: 4
        price:
          type: integer
          format: int32
          description: 価格
          example: 3800
        type:
          type: string
          description: プラン種別（SUBSCRIPTION / SINGLE）
          example: SUBSCRIPTION
        isRecommended:
          type: boolean
          description: おすすめバッジ
          example: true
      required:
      - beanQuantity
      - gramWeight
      - id
      - isRecommended
      - label
      - price
      - shopifyPlanId
      - type
...
  securitySchemes:
    Bearer:
      type: http
      description: JWTアクセストークン
      scheme: bearer
      bearerFormat: JWT